// web/admin/gql/schema.tsx
export type Maybe<T> = T | null | undefined
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] }

// Generated on 2021-10-04T01:13:08+05:30


/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  /** A location in a connection that can be used for resuming pagination. */
  Cursor: any
  /** A universally unique identifier as defined by [RFC 4122](https://tools.ietf.org/html/rfc4122). */
  UUID: string
  /** A point in time as described by the [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) standard. May or may not include a timezone. */
  Datetime: Date
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: { [key: string]: unknown }
  /** A floating point number that requires more precision than IEEE 754 binary 64 */
  BigFloat: any
  /** The `GeoJSON` scalar type represents GeoJSON values as specified by[RFC 7946](https://tools.ietf.org/html/rfc7946). */
  GeoJSON: GeometryObject
  Email: string
  /** Relaxed phone number validation */
  PhoneUs: any
  /** The day, does not include a time. */
  Date: any
  /** The exact time of day, does not include the date. May or may not have a timezone offset. */
  Time: string
  /** A JSON Web Token defined by [RFC 7519](https://tools.ietf.org/html/rfc7519) which securely represents claims between two parties. */
  JwtToken: string
}
// ...
export enum StatusT {
  Enabled = 'ENABLED',
  Disabled = 'DISABLED',
  Pending = 'PENDING',
}
// ...
export type VehicleOwnerIdFkeyInput = {
  /** The primary key(s) for `user` for the far side of the relationship. */
  connectById?: Maybe<UserUserPkeyConnect>
  /** The primary key(s) for `user` for the far side of the relationship. */
  connectByNodeId?: Maybe<UserNodeIdConnect>
  /** The primary key(s) for `user` for the far side of the relationship. */
  deleteById?: Maybe<UserUserPkeyDelete>
  /** The primary key(s) for `user` for the far side of the relationship. */
  deleteByNodeId?: Maybe<UserNodeIdDelete>
  /** The primary key(s) and patch data for `user` for the far side of the relationship. */
  updateById?: Maybe<UserOnVehicleForVehicleOwnerIdFkeyUsingUserPkeyUpdate>
  /** The primary key(s) and patch data for `user` for the far side of the relationship. */
  updateByNodeId?: Maybe<VehicleOnVehicleForVehicleOwnerIdFkeyNodeIdUpdate>
}
// ...
export type VehicleVehicleTypeIdFkeyInput = {
  /** The primary key(s) for `vehicleType` for the far side of the relationship. */
  connectById?: Maybe<VehicleTypeVehicleTypePkeyConnect>
  /** The primary key(s) for `vehicleType` for the far side of the relationship. */
  connectByName?: Maybe<VehicleTypeVehicleTypeNameUkeyConnect>
  /** The primary key(s) for `vehicleType` for the far side of the relationship. */
  connectByNodeId?: Maybe<VehicleTypeNodeIdConnect>
  /** The primary key(s) for `vehicleType` for the far side of the relationship. */
  deleteById?: Maybe<VehicleTypeVehicleTypePkeyDelete>
  /** The primary key(s) for `vehicleType` for the far side of the relationship. */
  deleteByName?: Maybe<VehicleTypeVehicleTypeNameUkeyDelete>
  /** The primary key(s) for `vehicleType` for the far side of the relationship. */
  deleteByNodeId?: Maybe<VehicleTypeNodeIdDelete>
  /** The primary key(s) and patch data for `vehicleType` for the far side of the relationship. */
  updateById?: Maybe<VehicleTypeOnVehicleForVehicleVehicleTypeIdFkeyUsingVehicleTypePkeyUpdate>
  /** The primary key(s) and patch data for `vehicleType` for the far side of the relationship. */
  updateByName?: Maybe<
    VehicleTypeOnVehicleForVehicleVehicleTypeIdFkeyUsingVehicleTypeNameUkeyUpdate
  >
  /** The primary key(s) and patch data for `vehicleType` for the far side of the relationship. */
  updateByNodeId?: Maybe<VehicleOnVehicleForVehicleVehicleTypeIdFkeyNodeIdUpdate>
  /** A `VehicleTypeInput` object that will be created and connected to this object. */
  create?: Maybe<VehicleVehicleTypeIdFkeyVehicleTypeCreateInput>
}
// ...
export type VehicleVehicleSizeIdFkeyInput = {
  /** The primary key(s) for `vehicleSize` for the far side of the relationship. */
  connectById?: Maybe<VehicleSizeVehicleSizePkeyConnect>
  /** The primary key(s) for `vehicleSize` for the far side of the relationship. */
  connectByNodeId?: Maybe<VehicleSizeNodeIdConnect>
  /** The primary key(s) for `vehicleSize` for the far side of the relationship. */
  deleteById?: Maybe<VehicleSizeVehicleSizePkeyDelete>
  /** The primary key(s) for `vehicleSize` for the far side of the relationship. */
  deleteByNodeId?: Maybe<VehicleSizeNodeIdDelete>
  /** The primary key(s) and patch data for `vehicleSize` for the far side of the relationship. */
  updateById?: Maybe<VehicleSizeOnVehicleForVehicleVehicleSizeIdFkeyUsingVehicleSizePkeyUpdate>
  /** The primary key(s) and patch data for `vehicleSize` for the far side of the relationship. */
  updateByNodeId?: Maybe<VehicleOnVehicleForVehicleVehicleSizeIdFkeyNodeIdUpdate>
  /** A `VehicleSizeInput` object that will be created and connected to this object. */
  create?: Maybe<VehicleVehicleSizeIdFkeyVehicleSizeCreateInput>
}
// ...
export type VehiclePatch = {
  id?: Maybe<Scalars['UUID']>
  ownerId?: Maybe<Scalars['UUID']>
  name?: Maybe<Scalars['String']>
  licensePlate?: Maybe<Scalars['String']>
  vehicleTypeId?: Maybe<Scalars['UUID']>
  vehicleSizeId?: Maybe<Scalars['UUID']>
  status?: Maybe<StatusT>
  createdAt?: Maybe<Scalars['Datetime']>
  updatedAt?: Maybe<Scalars['Datetime']>
  userToOwnerId?: Maybe<VehicleOwnerIdFkeyInput>
  vehicleTypeToVehicleTypeId?: Maybe<VehicleVehicleTypeIdFkeyInput>
  vehicleSizeToVehicleSizeId?: Maybe<VehicleVehicleSizeIdFkeyInput>
}
// ...
export type UpdateVehicleInput = {
  /** An arbitrary string value with no semantic meaning. Will be included in the payload verbatim. May be used to track mutations by the client. */
  clientMutationId?: Maybe<Scalars['String']>
  /** An object where the defined keys will be set on the `Vehicle` being updated. */
  patch: VehiclePatch
  id: Scalars['UUID']
}
// ...
export type ManageVehicleQuery = {
  vehicle?: Maybe<{
    id: string
    name: string
    status: StatusT
    licensePlate?: Maybe<string>
    vehicleTypeId: string
    vehicleSizeId: string
    ownerId: string
    vehicleType?: Maybe<{ id: string; name: string }>
    vehicleSize?: Maybe<{ id: string; name?: Maybe<string>; description?: Maybe<string> }>
    owner?: Maybe<{ id: string; name: string; email: string }>
  }>
}
// ...
export type UpdateVehicleMutation = {
  updateVehicle?: Maybe<{
    vehicle?: Maybe<{
      id: string
      name: string
      status: StatusT
      ownerId: string
      licensePlate?: Maybe<string>
      vehicleTypeId: string
    }>
  }>
}